+++ SleepModel/HEMSoundsContainerViewController.m
@@ -11,23 +11,29 @@
 #import "HEMSleepSoundService.h"
 #import "HEMAlarmService.h"
 #import "HEMDeviceService.h"
-#import "HEMSoundsSubNavPresenter.h"
+#import "HEMSoundsContentPresenter.h"
 #import "HEMActivityIndicatorView.h"
 #import "HEMSubNavigationView.h"
 #import "HEMAlarmListViewController.h"
 #import "HEMSleepSoundViewController.h"
+#import "HEMSensePairViewController.h"
+#import "HEMStyledNavigationViewController.h"
 #import "HEMMainStoryboard.h"
+#import "HEMOnboardingStoryboard.h"
+#import "HEMBaseController+Protected.h"
 
-@interface HEMSoundsContainerViewController()<HEMSoundSubNavDelegate>
+@interface HEMSoundsContainerViewController()<HEMSoundContentDelegate, HEMSensePairingDelegate>
 
 @property (weak, nonatomic) IBOutlet UIView *containerView;
 @property (weak, nonatomic) IBOutlet HEMActivityIndicatorView *activityIndicator;
 @property (weak, nonatomic) IBOutlet HEMSubNavigationView *subNav;
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *subNavHeightConstraint;
+@property (weak, nonatomic) IBOutlet UICollectionView *errorCollectionView;
 
 @property (nonatomic, strong) HEMSleepSoundService* sleepSoundsService;
 @property (nonatomic, strong) HEMAlarmService* alarmService;
 @property (nonatomic, strong) HEMDeviceService* deviceService;
+@property (nonatomic, weak) HEMSoundsContentPresenter* contentPresenter;
 
 @property (nonatomic, strong) HEMAlarmListViewController* alarmVC;
 @property (nonatomic, strong) HEMSleepSoundViewController* sleepSoundVC;
@@ -56,20 +62,39 @@
     [self setAlarmService:[HEMAlarmService new]];
     [self setDeviceService:[HEMDeviceService new]];
     
-    HEMSoundsSubNavPresenter* subNavPresenter
-        = [[HEMSoundsSubNavPresenter alloc] initWithSleepSoundService:[self sleepSoundsService]
-                                                         alarmService:[self alarmService]
-                                                        deviceService:[self deviceService]];
-    [subNavPresenter setDelegate:self];
-    [subNavPresenter bindWithActivityIndicator:[self activityIndicator]];
-    [subNavPresenter bindWithSubNavigationView:[self subNav]
+    HEMSoundsContentPresenter* presenter
+        = [[HEMSoundsContentPresenter alloc] initWithSleepSoundService:[self sleepSoundsService]
+                                                          alarmService:[self alarmService]
+                                                         deviceService:[self deviceService]];
+    [presenter setDelegate:self];
+    [presenter bindWithActivityIndicator:[self activityIndicator]];
+    [presenter bindWithSubNavigationView:[self subNav]
                           withHeightConstraint:[self subNavHeightConstraint]];
-    [self addPresenter:subNavPresenter];
+    [presenter bindWithErrorCollectionView:[self errorCollectionView]];
+    
+    [self setContentPresenter:presenter];
+    [self addPresenter:presenter];
 }
 
-#pragma mark - Sub Nav Delegate
+#pragma mark - Content Delegate
 
-- (void)loadAlarms:(BOOL)hasSensePaired {
+- (void)unloadContentControllersFrom:(HEMSoundsContentPresenter*)presenter {
+    UIViewController* currentVC = [[self childViewControllers] firstObject];
+    if (currentVC) {
+        [currentVC willMoveToParentViewController:nil];
+        [[currentVC view] removeFromSuperview];
+        [currentVC removeFromParentViewController];
+    }
+}
+
+- (void)pairWithSenseFrom:(HEMSoundsContentPresenter*)presenter {
+    HEMSensePairViewController *pairVC = (id)[HEMOnboardingStoryboard instantiateSensePairViewController];
+    [pairVC setDelegate:self];
+    UINavigationController *nav = [[HEMStyledNavigationViewController alloc] initWithRootViewController:pairVC];
+    [self presentViewController:nav animated:YES completion:nil];
+}
+
+- (void)loadAlarmsFrom:(__unused HEMSoundsContentPresenter *)presenter {
     DDLogVerbose(@"show alarms view");
     if (![self alarmVC]) {
         HEMAlarmListViewController* alarmList = [HEMMainStoryboard instantiateAlarmListViewController];
@@ -79,7 +104,7 @@
     [self showSoundViewOf:[self alarmVC]];
 }
 
-- (void)loadSleepSounds:(SENSleepSounds *)sleepSounds {
+- (void)loadSleepSounds:(SENSleepSounds *)sleepSounds from:(__unused HEMSoundsContentPresenter *)presenter {
     DDLogVerbose(@"show sleep sounds view");
     if (![self sleepSoundVC]) {
         HEMSleepSoundViewController* soundVC = [HEMMainStoryboard instantiateSleepSoundViewController];
@@ -118,6 +143,36 @@
                             [[self subNav] setUserInteractionEnabled:YES];
                         }];
     }
+}
+
+#pragma mark - HEMSensePairDelegate
+
+- (void)dismissModalAfterDelay:(BOOL)delay {
+    if (delay) {
+        NSTimeInterval delayInSeconds = 1.5f;
+        dispatch_time_t delay = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delayInSeconds * NSEC_PER_SEC));
+        dispatch_after(delay, dispatch_get_main_queue(), ^(void) {
+            [self dismissViewControllerAnimated:YES completion:nil];
+        });
+    } else {
+        [self dismissViewControllerAnimated:YES completion:nil];
+    }
+}
+
+- (void)didReturnWithSenseManager:(SENSenseManager*)senseManager {
+    BOOL paired = senseManager != nil;
+    if (paired) {
+        [[self contentPresenter] reload];
+    }
+    [self dismissModalAfterDelay:paired];
+}
+
+- (void)didPairSenseUsing:(SENSenseManager *)senseManager from:(UIViewController *)controller {
+    [self didReturnWithSenseManager:senseManager];
+}
+
+- (void)didSetupWiFiForPairedSense:(SENSenseManager *)senseManager from:(UIViewController *)controller {
+    [self didReturnWithSenseManager:senseManager];
 }
 
 @end